add progress bar loading
